Libft complete

+++ ft_memmove.c
@@ -14,8 +14,8 @@
 static unsigned char	*destbsrc(unsigned char *d, unsigned char *s,
 size_t len)
 {
-	size_t i;
-	
+	int	i;
+
 	i = len - 1;
 	while (i >= 0)
 	{
@@ -27,7 +27,7 @@ size_t len)
 
 void	*ft_memmove(void *dest, const void *src, size_t len)
 {
-	size_t			i;
+	int				i;
 	unsigned char	*d;
 	unsigned char	*s;
 
@@ -43,7 +43,7 @@ void	*ft_memmove(void *dest, const void *src, size_t len)
 	if (s > d && s - d < (int)len)
 	{
 		i = 0;
-		while (i < len)
+		while (i < (int)len)
 		{
 			d[i] = s[i];
 			i++;
